+++ frontend/src/Pages/sensores/ItemSensor.jsx
@@ -1,12 +1,38 @@
 import React from "react";
 import useAuth from "../../hooks/useAuth";
 import useSensores from "../../hooks/useSensores";
+import Swal from "sweetalert2";
+import { useNavigate } from "react-router-dom";
 
 const ItemSensor = ({ sensor }) => {
   const { auth } = useAuth();
   const { eliminarSensor } = useSensores();
+  const navigate = useNavigate();
+
   const handleDelete = () => {
-    eliminarSensor(sensor.idsensor);
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "Una vez eliminada, no podrás recuperar este sensor.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        eliminarSensor(sensor.idsensor);
+        Swal.fire(
+          "Eliminada",
+          "El sensor ha sido eliminado correctamente.",
+          "success"
+        );
+      }
+    });
+  };
+
+  const handleSelectSensor = () => {
+    navigate(`/sensores/editar/${sensor.idsensor}`);
   };
   return (
     <>
@@ -39,9 +65,12 @@ const ItemSensor = ({ sensor }) => {
             </div>
             {auth._cedula ? (
               <div className="text-center pt-3">
-                <a className="btn btn-info d-grid w-auto mx-2" href="">
+                <button
+                  onClick={handleSelectSensor}
+                  className="btn btn-info d-grid w-auto mx-2"
+                >
                   <i className="fa fa-pen mr-2"></i>Editar
-                </a>
+                </button>
                 <a
                   className="btn btn-danger d-grid w-auto mx-2"
                   onClick={handleDelete}
